requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-haystack==2.4.1
django-toolbelt==0.0.1
elasticsearch>=1.0.0,<2.0.0
gunicorn==19.4.3
psycopg2==2.6.1
static3==0.6.1
wsgiref==0.1.2
